backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pymongo==4.13.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
azure-ai-contentsafety==1.0.0 
motor==3.7.1